# Tensorboard callback modifications