--- a/keras/callbacks.py
+++ b/keras/callbacks.py
@@ -613,6 +613,8 @@ class TensorBoard(Callback):
             for histograms computation.
         write_images: whether to write model weights to visualize as
             image in TensorBoard.
+        write_batch_performance: whether to write training metrics on batch 
+            completion 
         embeddings_freq: frequency (in epochs) at which selected embedding
             layers will be saved.
         embeddings_layer_names: a list of names of layers to keep eye on. If
@@ -630,6 +632,7 @@ class TensorBoard(Callback):
                  write_graph=True,
                  write_grads=False,
                  write_images=False,
+                 write_batch_performance=False,
                  embeddings_freq=0,
                  embeddings_layer_names=None,
                  embeddings_metadata=None):
@@ -643,10 +646,12 @@ class TensorBoard(Callback):
         self.write_graph = write_graph
         self.write_grads = write_grads
         self.write_images = write_images
+        self.write_batch_performance = write_batch_performance
         self.embeddings_freq = embeddings_freq
         self.embeddings_layer_names = embeddings_layer_names
         self.embeddings_metadata = embeddings_metadata or {}
         self.batch_size = batch_size
+        self.epoch_current = 0
 
     def set_model(self, model):
         self.model = model
@@ -741,7 +746,7 @@ class TensorBoard(Callback):
 
     def on_epoch_end(self, epoch, logs=None):
         logs = logs or {}
-
+        self.epoch_current += 1
         if self.validation_data and self.histogram_freq:
             if epoch % self.histogram_freq == 0:
 
@@ -788,6 +793,23 @@ class TensorBoard(Callback):
 
     def on_train_end(self, _):
         self.writer.close()
+
+    def on_batch_end(self, batch, logs=None):
+        print('on batch end')
+        logs = logs or {}
+        print(logs)
+        #batch_size = logs.get('size', 0)
+        # best way is probably to shift everything to use self.seen (and track self.seen, of course)
+
+        for name, value in logs.items():
+            if name in ['batch','size']:
+                continue
+            summary = tf.Summary()
+            summary_value = summary.value.add()
+            summary.value.simple_value = value.item()
+            summary_value.tag = name
+            self.writer.add_summary(summary)
+        self.writer.flush()
 
 
 class ReduceLROnPlateau(Callback):