--- a/tests/keras/test_callbacks.py
+++ b/tests/keras/test_callbacks.py
@@ -347,7 +347,8 @@ def test_TensorBoard():
                   metrics=['accuracy'])
 
     tsb = callbacks.TensorBoard(log_dir=filepath, histogram_freq=1,
-                                write_images=True, write_grads=True,
+                                write_images=True, write_grads=True, 
+                                write_batch_performance=True,
                                 embeddings_freq=1,
                                 embeddings_layer_names=['dense_1'],
                                 batch_size=5)