This PR adds the option to write batch-level performance to Tensorboard. It's associated w/ this [thread](https://github.com/fchollet/keras/issues/6692). Apologies if this should go elsewhere or requires more discussion before creating a PR. Thanks