@sameermanek 2017-08-17
- This could modify existing behavior:

1. It adds an argument `write_batch_performance` to the Tensorboard callback constructor (defaults to False). It's currently in the middle of the argument list, near the (logically) related arguments, but this could change the behavior if any users are relying on positional rather than named arguments here. I think this is unlikely because there were already 9 named arguments, all of which have default values (and the usage examples I can find all use named arguments). If it makes more sense, I can move this to the end of the argument list, which would eliminate the behavior change. 

2. This also slightly changes the x-axis on the Tensorboard summaries to be the number of examples seen instead of the number of epochs

- This is being incorporated into the existing Tensorboard tests (in test_callbacks.py). 

Let me know if I should change any of these. Thanks