@bdwyer2 2017-08-24
@sameermanek thanks for doing this! Have you verified that the validation statistics are correctly reported with this PR?