@sameermanek 2017-08-25
@bdwyer2 Sorry, I don't understand -- what are the validation statistics? I ran all the tests locally and they succeeded (so those matched the Travis CI build tests) if that's what you're referring to. Thanks.