@ViaFerrata 2017-10-06
Question: Did you investigate the performance impact of logging after each batch? For RAM it shouldn't matter (even with >10 Mio. dataset like I use), but what about computation time?
If it is significant, one could think about an option to e.g. only log every hundredth batch.