@karantan 2017-10-11
Will this allow using `histogram_freq=1` option in TensorBoard callback when using `fit_generator`?