@sameermanek 2017-10-18
@ViaFerrata I ran a quick (totally non-scientific) test, and the performance impact was nearly negligible. I trained a network with and without the batch-level logging (alternating between with and without batch-level callback, 5 iterations each, 98 batches and 100 epochs each iteration)

Without batch-level logging: 142.4 seconds per iteration on average, sd of 7.1 seconds
With batch-level logging: 143.9 seconds per iteration, sd of 5.7 seconds

@karantan I'm not sure I follow -- this doesn't address any outstanding issues if that's what you're referring to.