@cdancette 2017-11-12
This PR would be very useful, any idea when it will be merged ?