@chipgarner 2017-12-21
Please! I am running @sameermanek 's version, Keras needs this.