@rafaspadilha 2018-01-15
This PR would be very useful!