@chipgarner 2018-01-15
This is working well for us on Ubuntu 16.04, Windows 10, and AWS servers using their Deep Learning AMI (Ubuntu) and modifying the one file. Needs one approved review!