@ahundt 2018-02-01
I've tested this and it is working well.