@Faur 2018-07-16
Bump. 

I think this is very useful, and it looks like most of the work has been done. Would really like to see this implemented, thank you!