@nandaw 2018-08-10
This would be really helpful for us working with huge datasets. Is there anything that I can do to help?