@fchollet 2018-09-12
I still think this is a good feature to add. Sorry for the delay in processing the PR.

@omalleyt12 : this should be your area of expertise. Could you please take it over?