@gabrieldemarmiesse 2018-10-01
This feature is now implemented and merged in master.